Change server folder structure
