+++ server/src/models/comment.py
@@ -1,7 +1,7 @@
 from pydantic import BaseModel, Field
 from typing import List, Any
 from datetime import datetime
-from objectid import ObjectId
+from .objectid import ObjectId
 
 class _CommentBase(BaseModel):
     id: ObjectId = Field(alias="_id")
