+++ server/src/routes/v1/__init__.py
@@ -1,0 +1,8 @@
+from fastapi import APIRouter
+
+router = APIRouter()
+
+
+@router.get("/")
+def index():
+    return {"msg": "You have successfully reached API v1"}
