--- server/src/table_names.py
@@ -1,6 +1,0 @@
-from enum import Enum
-
-
-class DbName(Enum):
-	COURSE = "courses"
-	USER = "users"
