+++ server/src/utils/json_encoder.py
@@ -1,0 +1,8 @@
+from json import JSONEncoder as BaseJSONEncoder
+
+
+class JSONEncoder(BaseJSONEncoder):
+    def default(self, o):
+        if isinstance(o, (ObjectId, DBRef)):
+            return str(o)
+        return json.JSONEncoder.default(self, o)
